Rewrote SimViewer as a SceneManager, changed styles and layout.
Styles and layout have been changed to support resizing the window and
having the render still take up the full content area below the ui bar.
The window can be resized, but not all the objects are back on the plane
as I'm trying to rework the three js code to allow adding specific
scene subjects to the scene independently. Next step is to get the
green bushy things back and then rework the Element buttons to place
animals on the plane again.