Figured out how to detect intersection of objects with mouse cursor
Took quite a bit of effort, as the header has to be accounted for when
calculating the most position and converting its coordinates to
normalized coordinates. The color or something needs to change when
the user moves their mouse over an object to help them see they are
active on a specific model.