Added getValue function helpers
The function solves a pretty irritating problem with javascript,
which is trying to access keys on null or undefined objects.
Instead of pulling in a large library like Lodash I wrote this
simpler function. It uses dot notation to get nested values from
an object, such as getValue('name.first', personObject).