Worked on functionality for maintaining a list of selected models.
This feels messy, but it works right now. The list is just stored in
an array and checked when you mouseover or click on things. The mouse
over check is happening as part of the mouseover highlight so that
the SceneManager knows what color to return to. Objects can specify
a color key as part of their userData and that is used to determine
original, selected, and highlighted colors.