Pulled capi related code out of App.js and into its own class.
CAPI is complex and confusing. I am still learning how to use this
integration. I think I see a path forward to building an example
of how authors and learners may use the platform. Next steps are
to allow the authors to specify the number of objects from SmartSparrow
instead of having it hard coded, and allowing them to set instructions
for learners based on those objects, such as "Select 3 Cedar trees".