Trying to get data from SmartSparrow.
What I learned is basically you need a callback for the Transporter if
you want it to give you data. The Transporter will tell your application
when the SmartSparrow platform is done setting up. I still do not know how
to get data out of the components in SmartSparrow. This is a key point
to learn to be able to make a 2 way integration.