Tried more things to get data I thought should be there
Spoke with Josef today (our sponsor) and after a few back and forth
discussions I came to realize what I'm trying to do is impossible
with the smart sparrow platform. I will be ignoring the built in
Smart Sparrow components going forward as those are not useful
for this type of project.